==> 02. Exploratory Data Analysis.ipynb <==
# Exploratory Data Analysis

In this notebook we'll do some exploratory data analysis over our dataset. However, since we don't have our features created yet, we cannot do much at this point. In addition, when we create them, we won't be able to extract many insights because of the nature of text-based features. For this reason, only a shallow analysis will be done at this point.

For the plots we have used `seaborn` and `altair`. `altair` is a package which allows us to plot graphics with a simple grammar as we would do in ggplot2 or Tableau. It also provides easy-to-make interactive plots. For further information please visit the project site: https://altair-viz.github.io/.

To install it, please type this command in the shell:

`! conda install -c conda-forge altair vega_datasets notebook vega`

import pandas as pd
import matplotlib.pyplot as plt
import pickle
# import seaborn as sns
# sns.set_style("whitegrid")
# import altair as alt
# alt.renderers.enable("notebook")

# Code for hiding seaborn warnings
import warnings
warnings.filterwarnings("ignore")

Loading the dataset:

df = pd.read_excel("data.xlsx")

df.to_csv("sail.csv", index = False)

df.head()

df.drop(["Remarks"], axis = 1, inplace = True)

df.Label.value_counts()

df.shape

df.isnull().sum()

df = df[pd.notnull(df['Tender Title'])]

df.isnull().sum()

df.shape

### Number of articles in each category

# bars = alt.Chart(df).mark_bar(size=50).encode(
#     x=alt.X("Label"),
#     y=alt.Y("count():Q", axis=alt.Axis(title='Number of examples')),
#     tooltip=[alt.Tooltip('count()', title='Number of examples'), 'Label'],
#     color='Label'

# )

# text = bars.mark_text(
#     align='center',
#     baseline='bottom',
# ).encode(
#     text='count()'
# )

# (bars + text).interactive().properties(
#     height=300, 
#     width=700,
#     title = "Number of examples in each category",
# )

### % of articles in each category

# df['id'] = 1
# df2 = pd.DataFrame(df.groupby('Classification').count()['id']).reset_index()

# bars = alt.Chart(df2).mark_bar(size=50).encode(
#     x=alt.X('Label'),
#     y=alt.Y('PercentOfTotal:Q', axis=alt.Axis(format='.0%', title='% of Clauses')),
#     color='Label'
# ).transform_window(
#     TotalArticles='sum(id)',
#     frame=[None, None]
# ).transform_calculate(
#     PercentOfTotal="datum.id / datum.TotalArticles"
# )

# text = bars.mark_text(
#     align='center',
#     baseline='bottom',
#     #dx=5  # Nudges text to right so it doesn't appear on top of the bar
# ).encode(
#     text=alt.Text('PercentOfTotal:Q', format='.1%')
# )

# (bars + text).interactive().properties(
#     height=300, 
#     width=700,
#     title = "% of Clauses in each category",
# )

### News length by category

Definition of news length field. Although there are special characters in the text (``\r, \n``), it will be useful as an approximation.

# df['Clause_length'] = df['Clause'].str.len()

# plt.figure(figsize=(12.8,6))
# sns.distplot(df['Clause_length']).set_title('Clause length distribution');

# df['Clause_length'].describe()

Let's remove from the 95% percentile onwards to better appreciate the histogram:

# quantile_95 = df['Clause_length'].quantile(0.95)
# df_95 = df[df['Clause_length'] < quantile_95]

# plt.figure(figsize=(12.8,6))
# sns.distplot(df_95['Clause_length']).set_title('Clause length distribution');

We can get the number of news articles with more than 10,000 characters:

# df_more10k = df[df['News_length'] > 10000]
# len(df_more10k)

Let's see one:

# df_more10k['Content'].iloc[0]

It's just a large news article.

Let's now plot a boxplot:

# plt.figure(figsize=(12.8,6))
# sns.boxplot(data=df, x='Label', y='Clause_length', width=.5);

Now, let's remove the larger documents for better comprehension:

# plt.figure(figsize=(12.8,6))
# sns.boxplot(data=df_95, x='Category', y='News_length');

We can see that, although the length distribution is different for every category, the difference is not too big. If we had way too different lengths between categories we would have a problem since the feature creation process may take into account counts of words. However, when creating the features with TF-IDF scoring, we will normalize the features just to avoid this.

At this point, we cannot do further Exploratory Data Analysis. We'll turn onto the **Feature Engineering** section.

We'll save the dataset:

with open('Contract_dataset.pickle', 'wb') as output:
    pickle.dump(df, output)
