==> 15. Sample Data.ipynb <==
# Sample Articles

We have to import:

* Trained models
* TF-IDF model

import pickle
import pandas as pd
import nltk
from nltk.corpus import stopwords
from nltk.tokenize import punkt
from nltk.corpus.reader import wordnet
from nltk.stem import WordNetLemmatizer
from sklearn.feature_extraction.text import TfidfVectorizer
import numpy as np

#### Trained models

path_models = "Models/"

# SVM
path_svm = path_models + 'best_mnbc.pickle'
with open(path_svm, 'rb') as data:
    mnb_model = pickle.load(data)

#### TF-IDF object

path_tfidf = "Pickles/tfidf.pickle"
with open(path_tfidf, 'rb') as data:
    tfidf = pickle.load(data)

#### Category mapping dictionary

category_codes = {
    'electronics': 1,
    'hardware': 2,
    'machine': 3,
    'none': 4,
    'raw_materials': 5,
    'skilled_manpower' : 6,
    'unskilled_manpower' : 7,
    'vehicle/equipment_hiring' : 8
}

#### Feature engineering workflow

punctuation_signs = list("?:!.,;")
stop_words = list(stopwords.words('english'))

def create_features_from_text(text):
    
    # Dataframe creation
    lemmatized_text_list = []
    df = pd.DataFrame(columns=['Content'])
    df.loc[0] = text
    df['Content_Parsed_1'] = df['Content'].str.replace("\r", " ")
    df['Content_Parsed_1'] = df['Content_Parsed_1'].str.replace("\n", " ")
    df['Content_Parsed_1'] = df['Content_Parsed_1'].str.replace("    ", " ")
    df['Content_Parsed_1'] = df['Content_Parsed_1'].str.replace('"', '')
    df['Content_Parsed_2'] = df['Content_Parsed_1'].str.lower()
    df['Content_Parsed_3'] = df['Content_Parsed_2']
    for punct_sign in punctuation_signs:
        df['Content_Parsed_3'] = df['Content_Parsed_3'].str.replace(punct_sign, '')
    df['Content_Parsed_4'] = df['Content_Parsed_3'].str.replace("'s", "")
    wordnet_lemmatizer = WordNetLemmatizer()
    lemmatized_list = []
    text = df.loc[0]['Content_Parsed_4']
    text_words = text.split(" ")
    for word in text_words:
        lemmatized_list.append(wordnet_lemmatizer.lemmatize(word, pos="v"))
    lemmatized_text = " ".join(lemmatized_list)    
    lemmatized_text_list.append(lemmatized_text)
    df['Content_Parsed_5'] = lemmatized_text_list
    df['Content_Parsed_6'] = df['Content_Parsed_5']
    for stop_word in stop_words:
        regex_stopword = r"\b" + stop_word + r"\b"
        df['Content_Parsed_6'] = df['Content_Parsed_6'].str.replace(regex_stopword, '')
    df = df['Content_Parsed_6']
#     df = df.rename(columns={'Content_Parsed_6': 'Content_Parsed'})
    
    # TF-IDF
    features = tfidf.transform(df).toarray()
    
    return features

function that tells us the category given the category code:

def get_category_name(category_id):
    for category, id_ in category_codes.items():    
        if id_ == category_id:
            return category

function that includes the whole process:

def predict_from_text(text):
    
    # Predict using the input model
    prediction_mnb = mnb_model.predict(create_features_from_text(text))[0]
    
    # Return result
    category_mnb = get_category_name(prediction_mnb)
    
#     print("The predicted category using the MNB model is %s." %(category_mnb) )
    return (category_mnb)

def predict_from_text_prob(text):
    
    # Predict using the input model
    prediction_mnb_proba = mnb_model.predict_proba(create_features_from_text(text))[0]
    
    
#     print("The conditional probability is: %a" %(prediction_mnb_proba.max()*100))
    return (prediction_mnb_proba.max()*100)

### Text introduction and prediction

Please introduce `text`:

text = """

Procurement of Rough Terrain Crane (100T Capacity) for FM(M)

"""

predict_from_text(text)

predict_from_text_prob(text)

# test = pd.read_csv("C:\\Users\\Ayushi.Goel\\Documents\\Work\\Contract\\Untitled Folder\\document.csv")

# test.head()

# test["predicted_class"] = test.apply(lambda row : predict_from_text(row["Clause"]), axis = 1) 

# test["predicted_class_prob"] = test.apply(lambda row : predict_from_text_prob(row["Clause"]), axis = 1) 

# test.predicted_class.value_counts()

# test.head()

# a = test.sort_values(['predicted_class_prob'],ascending=False).groupby(["predicted_class","predicted_class_prob"])

# a.first()

# test1 = test = pd.read_excel("C:\\Users\\Ayushi.Goel\\Documents\\Work\\Contract\\Sample Data\\sample2.xlsx")
b = pd.read_csv(r"C:\Users\Ayushi.Goel\Documents\Work\Contract\Elastic_ContractSearch_Result.csv")

b.Label.value_counts()

filtered_list = [    'Limitation_of_liability',
    'Liquidation_damages',
    'Indemnity',
    'Governing_Law',
    'Confidentiality',
    'Key_personnel',
    'Commencement_date',
    'Renewal_Term',
    'Payment_Term'
]

# filtered_list = ["Extra", "Delivery_Locations", "Attrition_Rate"]

a = b[b["Label"].isin(filtered_list)]
# b.drop(b[b['Label'] = filtered_list].index, inplace = True)

# d= {}
# with open(r"C:\Users\Ayushi.Goel\Documents\Work\Contract\Text Classification Code - Copy\elastic_output.txt") as f:
#     for line in f:
#         (key,val) = line.split()
#         dict[key] = val
        

# def dictToDf(dictionary):
#     newDict = {}
#     for x in dictionary:
#         for y in dictionary[x]:
#             newDict[y] = x

#     df = pd.DataFrame(list(newDict.items()), columns=['Values','Label'])
#     cols = df.columns.tolist()
#     cols = cols[-1:] + cols[:-1]
#     df = df[cols]
#     return df

# dictToDf()

a.Label.value_counts()

# a["Values"].replace(r"[\d]*[\.][\d]*", " ", regex = True, inplace = True)

a["Values"]

a["predicted_class"] = b.apply(lambda row : predict_from_text(row["Values"]), axis = 1) 
a["predicted_class_prob"] = b.apply(lambda row : predict_from_text_prob(row["Values"]), axis = 1) 

sor = a.sort_values(['predicted_class_prob'],ascending=False).groupby(["Label","predicted_class_prob"])

sor.first()

a1 = a.drop(a[a.Label != a.predicted_class].index)

sor = a.sort_values(['Label','predicted_class_prob'],ascending=True).groupby(["Label","predicted_class_prob"]).head(2)

sor1 = a1.sort_values(['Label','predicted_class_prob'],ascending=False).groupby(["Label","predicted_class_prob"]).head(2)

sor1

count=0
prev_label = ""
prev_value = ""
prev_prob = 0
sor3 = pd.DataFrame([],columns = ["Field","Values","Values_2","Confidence_Score(%)"])
for i,row in sor1.iterrows():
    prev_data = pd.DataFrame({'Field': row['Label'],'Values':prev_value,'Values_2':row['Values'],'Confidence_Score(%)':(row['predicted_class_prob']+prev_prob)/2}, index=[0])
    if count==1 and row['Label']==prev_label:
        sor3 = sor3.append(prev_data, ignore_index=True)  
        count=0
    elif count==0 and row['Label']==prev_label:
        continue
    elif count==0 and row['Label']!=prev_label:
        prev_label = row['Label']
        prev_prob = row['predicted_class_prob']
        prev_value = row['Values']
        count+=1
    elif count==1 and row['Label']!=prev_label:
        #sor3 = sor3.append(pd.DataFrame({'Label': row['Label'],'Values':row['Values'],'Values_2':"",'predicted_class_prob':row['predicted_class_prob']}, index=[0]), ignore_index=True)
        sor3 = sor3.append(pd.DataFrame({'Field': prev_label,'Values':prev_value,'Values_2':"",'Confidence_Score(%)':prev_prob}, index=[0]), ignore_index=True)

        count=0
        prev_label = row['Label']
        prev_prob = row['predicted_class_prob']
        prev_value = row['Values']
        count+=1

for x in set(np.array(a.Label))-set(np.array(a1.Label)):
    print (x)
    for i,row in sor.iterrows():
        
        prev_data = pd.DataFrame({'Field': row['Label'],'Values':prev_value,'Values_2':row['Values'],'Confidence_Score(%)':(row['predicted_class_prob']+prev_prob)/2}, index=[0])
    
        if count==1 and row['Label']==prev_label and row['Label']==x:
            sor3 = sor3.append(prev_data, ignore_index=True)  
            count=0
        elif count==0 and row['Label']==prev_label and row['Label']==x:
            continue
        elif count==0 and row['Label']!=prev_label and row['Label']==x:
            prev_label = row['Label']
            prev_prob = row['predicted_class_prob']
            prev_value = row['Values']
            count+=1
        elif count==1 and row['Label']!=prev_label and row['Label']==x:
            sor3 = sor3.append(pd.DataFrame({'Field': prev_label,'Values':prev_value,'Values_2':"",'Confidence_Score(%)':prev_prob}, index=[0]), ignore_index=True)

            count=0
            prev_label = row['Label']
            prev_prob = row['predicted_class_prob']
            prev_value = row['Values']
            count+=1

sor3

sor2 = a1.groupby(["Label"]).predicted_class_prob.nlargest(3)

sor2

sor1.predicted_class[sor1.ix[(sor1["predicted_class"]=="Limitation_of_liability") & (sor1["predicted_class_prob"]<55)].index]="None"

sor1.predicted_class[sor1.ix[(sor1["predicted_class"]=="Renewal_Term") & (sor1["predicted_class_prob"]<90)].index]="None"

sor1.predicted_class[sor1.ix[(sor1["predicted_class"]=="Key_personnel") & (sor1["predicted_class_prob"]<75)].index]="None"

sor1.predicted_class[sor1.ix[(sor1["predicted_class"]=="Indemnity") & (sor1["predicted_class_prob"]<85)].index]="None"

sor1.predicted_class[sor1.ix[(sor1["predicted_class"]=="Indemnity") & (sor1["predicted_class_prob"]<85)].index]="None"

sor1.predicted_class[sor1.ix[(sor1["predicted_class"]=="Governing_Law") & (sor1["predicted_class_prob"]<99)].index]="None"

sor1.predicted_class[sor1.ix[(sor1["predicted_class"]=="Confidentiality") & (sor1["predicted_class_prob"]<90)].index]="None"

sor1.predicted_class[sor1.ix[(sor1["predicted_class"]=="Commencement_date") & (sor1["predicted_class_prob"]<55)].index]="None"

sor1

df_final = sor1.drop(sor1[sor1.predicted_class == "None"].index)

df_final

df_final.drop("Label", axis = 1 , inplace = True)

df_final

df_final.rename(index=str,columns={"predicted_class":"Field",'predicted_class_prob':'Confidence_Score(%)'})

# b.predicted_class[b.ix[(b["predicted_class"]=="Limitation_of_Liability") & (b["predicted_class_prob"]<65)].index]="None"

# b.head(11)

# test1["predicted_class"] = test1.apply(lambda row : predict_from_text(row["Clause"]), axis = 1) 

# test1["predicted_class_prob"] = test1.apply(lambda row : predict_from_text_prob(row["Clause"]), axis = 1) 

# test1.predicted_class[test1.ix[(test1["predicted_class"]=="Limitation_of_Liability") & (test1["predicted_class_prob"]<65)].index]="None"

# test1.predicted_class[test1.ix[(test1["predicted_class"]=="Renewal_Term") & (test1["predicted_class_prob"]<65)].index]="None"

# test1.head(11)

# test1.head()

# for val, prob in zip(test1["predicted_class"], test1["predicted_class_prob"]):
#     print(val)
#     if val == "Limitation_of_Liability" and prob < 65:
#         test1.predicted_class = None  
#     elif val == "Renewal_Term" and prob < 70:
#         test1.predicted_class = None 
#     else:
#         test1.predicted_class = val

# test.head(11)

# test1.sort_values("predicted_class", ascending = True).groupby("predicted_class").head(11)

## Threshold determination

### Threshold
